+++ src/main/scala/edu/ie3/simona/model/thermal/ThermalGrid.scala
@@ -67,7 +67,7 @@ final case class ThermalGrid(
       house.zip(lastHpState.thermalGridState.houseState) match {
         case Some((thermalHouse, lastHouseState)) =>
           val (updatedHouseState, _) =
-            thermalHouse.determineState(
+            thermalHouse.updateState(
               relevantData,
               lastHouseState,
               lastHpState.ambientTemperature.getOrElse(
@@ -477,7 +477,7 @@ final case class ThermalGrid(
   ): (Option[ThermalHouseState], Option[ThermalThreshold], Power) = {
     (house, state.houseState) match {
       case (Some(thermalHouse), Some(lastHouseState)) =>
-        val (newState, threshold) = thermalHouse.determineState(
+        val (newState, threshold) = thermalHouse.updateState(
           relevantData,
           lastHouseState,
           lastAmbientTemperature,
@@ -490,7 +490,7 @@ final case class ThermalGrid(
           )
         ) {
           val (fullHouseState, maybeFullHouseThreshold) =
-            thermalHouse.determineState(
+            thermalHouse.updateState(
               relevantData,
               lastHouseState,
               lastAmbientTemperature,
@@ -581,8 +581,8 @@ final case class ThermalGrid(
     /* House will be left with no influx in all cases. Determine if and when a threshold is reached */
     val maybeUpdatedHouseState =
       house.zip(lastThermalGridState.houseState).map {
-        case (house, houseState) =>
-          house.determineState(
+        case (thermalHouse, houseState) =>
+          thermalHouse.updateState(
             relevantData,
             houseState,
             lastAmbientTemperature,
@@ -679,7 +679,7 @@ final case class ThermalGrid(
           )
         ),
       )
-      val revisedHouseState = thermalHouse.determineState(
+      val revisedHouseState = thermalHouse.updateState(
         relevantData,
         formerHouseState.getOrElse(
           throw new InconsistentStateException(