+++ src/main/scala/edu/ie3/simona/model/thermal/ThermalHouse.scala
@@ -16,7 +16,7 @@ import edu.ie3.simona.model.participant.HpModel.HpRelevantData
 import edu.ie3.simona.model.thermal.ThermalGrid.ThermalEnergyDemand
 import edu.ie3.simona.model.thermal.ThermalHouse.ThermalHouseThreshold.{
   HouseTemperatureLowerBoundaryReached,
-  HouseTemperatureUpperBoundaryReached,
+  HouseTemperatureTargetOrUpperBoundaryReached,
 }
 import edu.ie3.simona.model.thermal.ThermalHouse.{
   ThermalHouseState,
@@ -238,7 +238,7 @@ final case class ThermalHouse(
     * @return
     *   Updated state and the tick in which the next threshold is reached
     */
-  def determineState(
+  def updateState(
       relevantData: HpRelevantData,
       state: ThermalHouseState,
       lastAmbientTemperature: Temperature,
@@ -319,7 +319,7 @@ final case class ThermalHouse(
         upperBoundaryTemperature,
         innerTemperature,
         resultingQDot,
-      ).map(HouseTemperatureUpperBoundaryReached)
+      ).map(HouseTemperatureTargetOrUpperBoundaryReached)
     } else {
       /* House is in perfect balance */
       None
@@ -403,7 +403,7 @@ object ThermalHouse {
     final case class HouseTemperatureLowerBoundaryReached(
         override val tick: Long
     ) extends ThermalThreshold
-    final case class HouseTemperatureUpperBoundaryReached(
+    final case class HouseTemperatureTargetOrUpperBoundaryReached(
         override val tick: Long
     ) extends ThermalThreshold
   }